+++ indispensable/src/commonMain/kotlin/at/asitplus/signum/indispensable/MessageAuthenticationCode.kt
@@ -4,8 +4,10 @@ import at.asitplus.signum.indispensable.asn1.*
 import at.asitplus.signum.indispensable.asn1.encoding.Asn1
 import at.asitplus.signum.indispensable.asn1.encoding.Asn1.Null
 import at.asitplus.signum.indispensable.asn1.encoding.readNull
+import at.asitplus.signum.indispensable.asn1.serialization.Asn1Serializer
 import at.asitplus.signum.indispensable.misc.BitLength
 import at.asitplus.signum.indispensable.misc.bit
+import kotlinx.serialization.Serializable
 
 sealed interface MessageAuthenticationCode : DataIntegrityAlgorithm {
     /** output size of MAC */
@@ -41,6 +43,7 @@ interface SpecializedMessageAuthenticationCode : SpecializedDataIntegrityAlgorit
 /**
  * RFC 2104 HMAC
  */
+@Serializable(with = HMAC.Companion::class)
 enum class HMAC(val digest: Digest, override val oid: ObjectIdentifier) : MessageAuthenticationCode, Identifiable,
     Asn1Encodable<Asn1Sequence> {
     SHA1(Digest.SHA1, KnownOIDs.hmacWithSHA1),
@@ -57,7 +60,7 @@ enum class HMAC(val digest: Digest, override val oid: ObjectIdentifier) : Messag
     }
 
 
-    companion object : Asn1Decodable<Asn1Sequence, HMAC> {
+    companion object : Asn1Decodable<Asn1Sequence, HMAC>, Asn1Serializer<Asn1Sequence, HMAC> {
 
         fun byOID(oid: ObjectIdentifier): HMAC? = entries.find { it.oid == oid }